Structural edits made to all files before filling in with code
